reorg and start to get user data
