+++ Overview.py
@@ -57,128 +57,10 @@ with st.expander("TODO list"):
     )
 
 
-df = utils.load_labeled_program_data()
-df["Date"] = pd.to_datetime(df["Date"])
 
 
-def get_program_chart_data(df, metric, num_programs, agg_method, date_range, exclude_solana):
-    if date_range == "All dates":
-        chart_df = df.copy()
-    elif date_range == "Year to Date":
-        chart_df = df.copy()[df.Date >= "2022-01-01"]
-    else:
-        chart_df = df.copy()[df.Date >= (datetime.datetime.today() - pd.Timedelta(date_range))]
-
-    if exclude_solana:
-        chart_df = chart_df[chart_df.LABEL != "solana"]
-    program_ids = (
-        chart_df.groupby("PROGRAM_ID")
-        .agg({metric: agg_method})
-        .sort_values(by=metric, ascending=False)
-        .iloc[:num_programs]
-        .index
-    )
-    chart_df = (
-        chart_df[chart_df.PROGRAM_ID.isin(program_ids)]
-        .sort_values(by=["Date", metric], ascending=False)
-        .reset_index()
-    )
-
-    return chart_df
-
-
-agg_method_dict = {
-    "mean": "Average within date range",
-    "sum": "Total within date range",
-    "max": "Highest daily usage during date range",
-}
-metric_dict = {"SIGNERS": "Number of signers", "TX_COUNT": "Transaction Count"}
-
-st.subheader("Programs")
-c1, c2, c3, c4 = st.columns(4)
-metric = c1.radio(
-    "Choose a metric",
-    metric_dict.keys(),
-    format_func=lambda x: metric_dict[x],
-    horizontal=True,
-    key="program_metric",
-)
-num_programs = c2.slider("Number of Programs", 1, 25, 10, key="program_slider")
-agg_method = c3.radio(
-    "Choose an aggregtion method",
-    agg_method_dict.keys(),
-    format_func=lambda x: agg_method_dict[x],
-    horizontal=True,
-    key="program_agg_method",
-)
-exclude_solana = c4.checkbox("Exclude Solana System Programs?", key="program_exclude_solana")
-date_range = st.radio(
-    "Choose a date range:",
-    [  # #TODO: not doing more dates until more data is queried
-        # "All dates",
-        # "Year to Date",
-        # "180d",
-        "90d",
-        "60d",
-        "30d",
-        "14d",
-        "7d",
-    ],
-    horizontal=True,
-    key="program_date_range",
-)
-
-chart_df = get_program_chart_data(df, metric, num_programs, agg_method, date_range, exclude_solana)
-chart_df["Name"] = chart_df.apply(
-    utils.apply_program_name,
-    axis=1,
-)
-
-chart = (
-    alt.Chart(chart_df)
-    .mark_line()
-    .encode(
-        x=alt.X("yearmonthdatehours(Date)", title=None),
-        y=alt.Y(metric, title=metric_dict[metric], sort="-y"),
-        color=alt.Color(
-            "Name",
-            title="Program Name",
-            scale=alt.Scale(scheme="turbo"),
-            sort=alt.EncodingSortField(metric, op=agg_method, order="descending"),
-        ),
-        tooltip=[
-            alt.Tooltip("yearmonthdatehours(Date)", title="Date"),
-            alt.Tooltip("TX_COUNT", title=metric_dict["TX_COUNT"], format=","),
-            alt.Tooltip("SIGNERS", title=metric_dict["SIGNERS"], format=","),
-            alt.Tooltip("PROGRAM_ID", title="Program ID"),
-            alt.Tooltip("LABEL_TYPE", title="Label Type"),
-            alt.Tooltip("LABEL_SUBTYPE", title="Label Subtype"),
-            alt.Tooltip("LABEL", title="Label"),
-            alt.Tooltip("ADDRESS_NAME", title="Address Name"),
-        ],
-    )
-    .properties(height=800, width=800)
-    .interactive()
-)
-st.altair_chart(chart, use_container_width=True)
-with st.expander("View and Download Data Table"):
-    st.write(chart_df)
-    st.download_button(
-        "Press to Download",
-        chart_df.to_csv().encode("utf-8"),
-        f"program_ids_top{num_programs}_{agg_method}_{date_range.replace(' ', '')}_{metric}.csv",
-        "text/csv",
-        key="download-program-ids",
-    )
-
-# #%ODO: melt, multiline tooltip
-# df = df.melt(
-#     id_vars=["Date", "CREATOR", "LABEL_TYPE", "LABEL_SUBTYPE", "LABEL", "ADDRESS_NAME"]
-# )
-# df
-
-
-st.header("Programs by date")
+st.header("Programs")
+st.write('See the [In Depth](In_Depth) page for more details.')
 weekly_program_data = utils.load_weekly_program_data()
 weekly_new_program_data = utils.load_weekly_new_program_data()
