--- pages/1_Developer.py
@@ -1,4 +1,0 @@
-import streamlit as st
-
-st.header("Developer")
-st.write("TODO: metrics of success stuff, comparison with eth blockspace")
