+++ pages/2_Developer.py
@@ -1,0 +1,24 @@
+import streamlit as st
+import altair as alt
+from PIL import Image
+
+alt.data_transformers.disable_max_rows()
+image = Image.open("assets/images/spire_background.png")
+
+st.set_page_config(
+    page_title="Spire FYI",
+    page_icon=image,
+    layout="wide",
+)
+c1, c2 = st.columns([1, 3])
+
+c2.header("Spire FYI")
+c2.caption("A viewpoint above Solana data")
+c1.image(
+    image,
+    width=200,
+)
+st.write("---")
+
+st.header("Developer")
+st.write("TODO: metrics of success stuff, comparison with eth blockspace")
