--- pages/2_Trending.py
@@ -1,4 +1,0 @@
-import streamlit as st
-
-st.header("Trending")
-st.write("TODO: Popular programs?")
